fix(api): clamp client pagination and drop unchecked offset math
Several list endpoints took page/per_page/limit straight from the query string
and used them without an upper bound:

  * /api2/search/ computed (page - 1) * per_page in i32, so a large page or
    page size overflowed - a panic in debug/test builds and a wrapped offset in
    release.
  * /api/v2.1/activities/ passed per_page to SQL as the LIMIT, so
    per_page=4294967295 pulled the entire filtered activity table into memory,
    and the same multiplication could overflow u32.
  * the tag endpoints left start/limit unbounded, and the file browser left
    page unbounded.
  * trash offsets multiplied page * per_page in u32.

Clamp each at the boundary (search 100/page 10k - 100 is what the iOS client
asks for, activities 100, tags 1000 which is what the mobile clients ask for,
browser page 10k), widen the trash math to u64, and compute the search offset
in i64 with saturating arithmetic. The repo dir endpoint stays unpaginated on
purpose: upstream /dir/ has no pagination either and the Android client
depends on a single full response.

Add tests for the clamped extremes.

+++ server/src/fs/core/trash.rs
@@ -309,7 +309,9 @@ pub async fn list_trash2(
 ) -> Result<TrashListResult, AppError> {
     let page = page.max(1);
     let per_page = per_page.clamp(1, 100);
-    let offset = ((page - 1) * per_page) as u64;
+    // Widen before multiplying: `page * per_page` in `u32` would overflow for
+    // large page numbers (panic in debug builds, wrapped offset in release).
+    let offset = u64::from(page - 1) * u64::from(per_page);
 
     // Count total
     let total_count = repos.file_trash.count_by_repo(repo_id).await?;
@@ -344,7 +346,9 @@ pub async fn search_trash(
 ) -> Result<TrashListResult, AppError> {
     let page = page.max(1);
     let per_page = per_page.clamp(1, 100);
-    let offset = ((page - 1) * per_page) as u64;
+    // Widen before multiplying: `page * per_page` in `u32` would overflow for
+    // large page numbers (panic in debug builds, wrapped offset in release).
+    let offset = u64::from(page - 1) * u64::from(per_page);
 
     let condition = build_trash_condition(
         Condition::all().add(file_trash::Column::RepoId.eq(repo_id.to_owned())),
@@ -776,7 +780,9 @@ pub async fn list_trash_for_user(
 ) -> Result<TrashListResult, AppError> {
     let page = page.max(1);
     let per_page = per_page.clamp(1, 100);
-    let offset = ((page - 1) * per_page) as u64;
+    // Widen before multiplying: `page * per_page` in `u32` would overflow for
+    // large page numbers (panic in debug builds, wrapped offset in release).
+    let offset = u64::from(page - 1) * u64::from(per_page);
 
     let repo_ids = gather_user_repo_ids(repos, user_id).await?;
 
@@ -830,7 +836,9 @@ pub async fn search_trash_for_user(
 ) -> Result<TrashListResult, AppError> {
     let page = page.max(1);
     let per_page = per_page.clamp(1, 100);
-    let offset = ((page - 1) * per_page) as u64;
+    // Widen before multiplying: `page * per_page` in `u32` would overflow for
+    // large page numbers (panic in debug builds, wrapped offset in release).
+    let offset = u64::from(page - 1) * u64::from(per_page);
 
     let repo_ids = gather_user_repo_ids(repos, user_id).await?;
